+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GtProcessDock: Dummy process elements are no longer mutable - #162
 - Examples search path was set to a specific 'examples' dir - #168
 - Fixed a crash when zooming in on a graphics scene containing a dense grid - #179
+- Fixed bug in dynamic type registration in meta object system - #248
 
 ### Changed
 - Saving a project was moved to a separate thread - #127